Extract tls-configuration.go from infra_test.go
