+++ tests/BUILD.bazel
@@ -181,7 +181,6 @@ go_test(
         "//pkg/util/hardware:go_default_library",
         "//pkg/util/migrations:go_default_library",
         "//pkg/util/net/dns:go_default_library",
-        "//pkg/util/tls:go_default_library",
         "//pkg/virt-api:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-controller/leaderelectionconfig:go_default_library",
