+++ tests/infrastructure/BUILD.bazel
@@ -7,10 +7,13 @@ go_library(
         "crds.go",
         "downward-metrics.go",
         "k8s-client-changes.go",
+        "tls-configuration.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/infrastructure",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/util/tls:go_default_library",
+        "//pkg/virt-config:go_default_library",
         "//pkg/virt-operator/resource/generate/components:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
@@ -19,6 +22,7 @@ go_library(
         "//tests/containerdisk:go_default_library",
         "//tests/decorators:go_default_library",
         "//tests/flags:go_default_library",
+        "//tests/framework/checks:go_default_library",
         "//tests/framework/kubevirt:go_default_library",
         "//tests/framework/matcher:go_default_library",
         "//tests/libinfra:go_default_library",
@@ -29,6 +33,7 @@ go_library(
         "//tests/util:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1:go_default_library",
         "//vendor/k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
